+++ bin/utils/line_helper.dart
@@ -1,8 +1,0 @@
-class LineHelper {
-  PreviousLine previousLine;
-  int calcLines;
-
-  LineHelper({this.previousLine = PreviousLine.notSet, this.calcLines = 0});
-}
-
-enum PreviousLine { notSet, refund }