+++ bin/utils/printing_helper.dart
@@ -4,17 +4,6 @@ import 'package:args/args.dart';
 import 'package:path/path.dart';
 import 'package:yaml/yaml.dart';
 
-/// Prints selected values.
-void printSelectedValues(String? selectedTextFile, String selectedHtmlFile,
-    String selectedStore, String exportFilesPath, String exportFilesFormat) {
-  print('Selected values:'
-      '${selectedTextFile != null ? '\n- Path to the cash receipt:\t\t\t\t$selectedTextFile' : ''}'
-      '\n- Path to the EAN products file:\t\t\t$selectedHtmlFile'
-      '\n- Selected store:\t\t\t\t\t$selectedStore'
-      '\n- Path where to save the output files:\t\t\t$exportFilesPath'
-      '\n- In which file format the output files will be saved:\t$exportFilesFormat\n');
-}
-
 /// Prints some basic info.
 Future<void> printBasicInfo(ArgParser parser) async {
   String pathToYaml =