+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.0.0
+## 2.0.0
 
-- Initial version.
+- Update the whole codebase to use [kassakuitti package](https://pub.dev/packages/kassakuitti).
+- Remove old unnecessary files.
+- Change Excel (xlsx) to the default export file format.
+
+## 0.10.0 - 1.0.2
+
+- Basic logic for S-kaupat & K-ruoka